added new rms norm extension layers, hardware not passing tb
