--- machop/chop/passes/graph/transforms/quantize/quantized_modules/rms_norm2d.py
@@ -1,22 +1,0 @@
-import logging
-
-import torch
-import torch.nn as nn
-from torch import Tensor
-
-from ..quantizers.integer import integer_floor_quantizer
-
-
-logger = logging.getLogger(__file__)
-logger.setLevel(logging.INFO)
-
-
-class _RMSNormBase(nn.Module):
-    def __init__(
-        self,
-        normalized_shape,
-    ) -> None:
-        super().__init__()
-
-    def forward(self, x: Tensor):
-        pass
